# perf(runtime/fs): optimize readFile by using a single large buffer