This avoids allocating N buffers when reading entire files and copying when concatenating them.

## Benchmarks

```
# Before
❯ deno run -A ./cli/bench/deno_common.js
read_128k_sync:      	n = 50000, dt = 1.941s, r = 25760/s, t = 38820ns/op
read_128k:           	n = 50000, dt = 9.613s, r = 5201/s, t = 192259ns/op

# After
❯ ./target/release/deno run -A ./cli/bench/deno_common.js
read_128k_sync:      	n = 50000, dt = 1.746s, r = 28637/s, t = 34920ns/op
read_128k:           	n = 50000, dt = 8.865s, r = 5640/s, t = 177300ns/op
```

## Notes

- The improvements to `read_128k_sync` are somewhat reduced here by the extra cwd lookups caused by the `stat`, so https://github.com/denoland/deno/pull/12056 will work hand in hand with this change to improve file reads
- This operates under the assumption that files won't change (truncated or extended) whilst being read, which isn't guaranteed but seems fair IMO
- Also simplifies implementation of `readTextFile` / `readTextFileSync`