@AaronO 2021-09-13
> Not in favor of this. File reads area not atomic. We can't start optimizing at the cost of correctness.

My point is more that there's no sane use-case to write & read-all at the same time when you don't control the rate of reading and it's uncommon and IMO would be mostly fine as is.

But as I mentioned we can make this more robust:
> Files being truncated should already be handled, files being extended could be handled, either by allocating an extra byte in the buffer and seeing if that is set or by doing an extra read.

essentially the current implementation would be a file path for an unchanged file and then it would fallback to the slowpath if the file was extended, so that wouldn't hurt or change correctness