@piscisaureus 2021-09-15
FWIW, node.js does exactly this. https://github.com/nodejs/node/blob/bbd4c6eee90ef600022c72316ba82d37a2ae16bc/lib/fs.js#L326-L349