@AaronO 2021-09-16
I still think that it would be sane/fair to not read beyond the stat'd size, I can't imagine a sane use-case for it but I implemented the slowpath fallback anyway.

@bartlomieju @lucacasonato That should address your concerns